requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
tiktoken
unstructured
python-docx
python-dotenv
